0.16.1: split run_wasm_gc + replay_cmd into focused submodules, two regression smokes
Three review followups before merge:

API cleanup
- `cmd_run_wasm_gc_with_mode_capturing` is gone; nobody was calling
  it with `Some(out_json)` after `try_run_wasm_gc` landed.
  `cmd_run_wasm_gc_with_mode` now wraps `try_run_wasm_gc` directly.
- `RunOutcome.replay_remaining` removed (computed but never read —
  the unconsumed-trailer check is enforced inside
  `try_run_wasm_gc` and surfaces as `Err(String)` instead).
- Stale doc comment about a "replayer that uses this capturing
  variant" deleted.

Module split
Following the file-size review point:

  run_wasm_gc.rs (2783 → 470 lines)
    + run_wasm_gc/decode.rs (602 lines): type-driven Val ↔ JsonValue
      decoders, `read_struct` over the wasmtime GC API, marker
      helpers, `encode_entry_args_for_wasm_gc`, LM transport bridges.
    + run_wasm_gc/imports.rs (1711 lines): `dispatch_aver_import`
      (the giant per-effect arms), record/replay glue
      (`try_replay`, `record_effect_if_recording`, `json_*`
      builders), all `host_*_make` factories, http_simple/body
      dispatch, host print + LM helpers.

  replay_cmd.rs (842 → 487 lines)
    + replay_cmd/backends.rs (344 lines): `BackendReplayOutcome`,
      `build_replay_result`, the three `run_<x>_replay()` functions
      (VM, wasm-gc, self-host) — exactly the surface the review
      asked be carved off so each backend's mechanics live next to
      its outcome shape.

Visibility tweak: `RunWasmGcHost` + its fields are `pub(super)` so
both submodules see them. Cross-module calls (decode ↔ imports) go
through `super::*` paths.

Two regression smokes
- `wasm_gc_replay_args_mismatch_warns_without_check_args_and_fails_with`:
  structurally swap the args of a recorded `Console.print` via
  `parse_session_recording → mutate → session_recording_to_string_pretty`,
  replay without `--check-args` (must succeed but surface
  `N effect(s) had different args`), then with `--check-args` (must
  fail). Catches the previous `args_diffs: 0` hardcode regression
  if it ever sneaks back.
- `wasm_gc_batch_replay_keeps_going_past_a_bad_recording`:
  `aver replay --wasm-gc <dir>` over a directory with one valid
  recording and one structurally-broken JSON; the runner must
  produce output for both files instead of `process::exit`-ing on
  the bad one. Locks in the no-`exit` contract from the review.

Plus the trailing-effect smoke (`wasm_gc_replay_rejects_recording_with_trailing_effect`)
now uses `parse_session_recording` + `EffectRecord` push +
`session_recording_to_string_pretty` instead of fragile substring
slicing — survives any future change to the pretty-printer's
formatting.

Test results
- `cargo test --release --features wasm`: every suite ok.
- `cargo test --release --features wasm --test
  wasm_gc_record_replay_spec`: 8/8 pass (the 5 from phase 4d, the
  trailing-effect smoke from the review fix, the two new ones
  added here).
- `cargo clippy --features wasm -- -D warnings` and the no-wasm
  invocation: clean.
