+++ src/main/replay_cmd.rs
@@ -1,21 +1,22 @@
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::process::{self, Command};
+use std::process;
 
 use colored::Colorize;
 
 use aver::ast::TopLevel;
-use aver::nan_value::{NanValue, NanValueConvert};
-use aver::replay::{
-    JsonValue, RecordedOutcome, SessionRecording, first_diff_path, format_json, json_to_value,
-    parse_session_recording, value_to_json,
-};
-use aver::value::{Value, aver_repr, list_to_vec};
-use aver::vm;
+use aver::replay::{JsonValue, SessionRecording, json_to_value, parse_session_recording};
+use aver::value::{Value, list_to_vec};
 
-use crate::commands::find_self_host_binary;
-use crate::shared::{apply_runtime_policy_to_vm, parse_file, read_file};
+use crate::shared::{parse_file, read_file};
 use aver::tty_render::render_tty;
+
+#[path = "replay_cmd/backends.rs"]
+mod backends;
+
+use backends::{build_replay_result, run_self_host_replay, run_vm_replay};
+#[cfg(feature = "wasm")]
+use backends::{find_fn_line_in_file, run_wasm_gc_replay};
 
 fn collect_recording_files_from_dir(dir: &Path, files: &mut Vec<PathBuf>) -> Result<(), String> {
     let entries = fs::read_dir(dir)
@@ -74,21 +75,6 @@ pub(super) fn decode_entry_args(input: &JsonValue) -> Result<Vec<Value>, String>
     }
 }
 
-fn decode_self_host_guest_args(input: &JsonValue) -> Result<Vec<String>, String> {
-    decode_entry_args(input)?
-        .into_iter()
-        .enumerate()
-        .map(|(idx, value)| match value {
-            Value::Str(s) => Ok(s),
-            other => Err(format!(
-                "Self-host replay expects guest input as List<String>; item {} was {}",
-                idx,
-                aver_repr(&other)
-            )),
-        })
-        .collect()
-}
-
 fn resolve_replay_module_root(path: &Path, recording: &SessionRecording) -> String {
     let module_root = Path::new(&recording.module_root);
     if module_root.is_absolute() {
@@ -111,7 +97,10 @@ fn resolve_replay_module_root(path: &Path, recording: &SessionRecording) -> Stri
     recording.module_root.clone()
 }
 
-fn resolve_replay_program_file(recording: &SessionRecording, module_root: &str) -> String {
+pub(super) fn resolve_replay_program_file(
+    recording: &SessionRecording,
+    module_root: &str,
+) -> String {
     let program_file = Path::new(&recording.program_file);
     if program_file.is_absolute() {
         return recording.program_file.clone();
@@ -130,7 +119,7 @@ fn resolve_replay_program_file(recording: &SessionRecording, module_root: &str) 
 }
 
 /// Find the line number of a key in raw JSON text (1-based, 0 = not found).
-fn find_json_line(raw: &str, key: &str) -> usize {
+pub(super) fn find_json_line(raw: &str, key: &str) -> usize {
     let pattern = format!("\"{}\"", key);
     for (idx, line) in raw.lines().enumerate() {
         if line.trim_start().starts_with(&pattern) {
@@ -140,7 +129,7 @@ fn find_json_line(raw: &str, key: &str) -> usize {
     0
 }
 
-fn find_fn_line(items: &[TopLevel], name: &str) -> usize {
+pub(super) fn find_fn_line(items: &[TopLevel], name: &str) -> usize {
     for item in items {
         if let TopLevel::FnDef(fd) = item
             && fd.name == name
@@ -151,96 +140,10 @@ fn find_fn_line(items: &[TopLevel], name: &str) -> usize {
     1
 }
 
-/// Backend-agnostic shape every replay backend reports back. Lets
-/// `replay_recording_file_*` functions stay focused on the
-/// per-backend mechanics (compile, run, parse subprocess output)
-/// while `build_replay_result` owns the `RecordedOutcome` ↔
-/// `recording.output` comparison and the `ReplayResult` shape the
-/// CLI renderer reads.
-pub(super) struct BackendReplayOutcome {
-    /// What the live re-run produced — JSON for normal returns, a
-    /// runtime-error string for traps / unhandled errors.
-    pub actual: RecordedOutcome,
-    /// `replay_pos` after the entry fn returned (`replay_progress.0`).
-    pub effects_consumed: usize,
-    /// `replay_effects.len()` from the recording (`replay_progress.1`).
-    pub effects_total: usize,
-    /// Soft-warning counter: `replay_effect` calls whose args didn't
-    /// match the recorded args. Only meaningful when `--check-args`
-    /// is OFF — with the flag, mismatches are already hard errors
-    /// inside `EffectReplayState::replay_effect`.
-    pub args_diff_count: usize,
-}
-
-/// Build the renderer-facing `ReplayResult` from a single backend's
-/// outcome. Centralises the `actual == recording.output` comparison
-/// plus the `output_diff` / `recording_output_line` plumbing so each
-/// backend doesn't reinvent them and a stray `args_diffs: 0`
-/// hardcode can't sneak back in.
-fn build_replay_result(
-    path: &Path,
-    raw: &str,
-    recording: &SessionRecording,
-    replay_program_file: String,
-    entry_line: usize,
-    outcome: Result<BackendReplayOutcome, String>,
-) -> ReplayResult {
-    let recording_output_line = find_json_line(raw, "output");
-    match outcome {
-        Ok(out) => {
-            let matched = out.actual == recording.output;
-            let output_diff = if !matched {
-                build_output_diff(&recording.output, &out.actual)
-            } else {
-                None
-            };
-            ReplayResult {
-                recording_path: path.display().to_string(),
-                program_file: replay_program_file,
-                entry_fn: recording.entry_fn.clone(),
-                entry_line,
-                matched,
-                effects_consumed: out.effects_consumed,
-                effects_total: out.effects_total,
-                error: None,
-                output_diff,
-                args_diffs: out.args_diff_count,
-                recording_output_line,
-            }
-        }
-        Err(e) => ReplayResult {
-            recording_path: path.display().to_string(),
-            program_file: replay_program_file,
-            entry_fn: recording.entry_fn.clone(),
-            entry_line,
-            matched: false,
-            effects_consumed: 0,
-            effects_total: recording.effects.len(),
-            error: Some(ReplayError::Generic(e)),
-            output_diff: None,
-            args_diffs: 0,
-            recording_output_line,
-        },
-    }
-}
-
-fn build_output_diff(
-    expected: &RecordedOutcome,
-    actual: &RecordedOutcome,
-) -> Option<(String, String, Option<String>)> {
-    match (expected, actual) {
-        (RecordedOutcome::Value(exp), RecordedOutcome::Value(got)) => {
-            let diff_path = first_diff_path(exp, got).map(|p| p.to_string());
-            Some((format_json(exp), format_json(got), diff_path))
-        }
-        (RecordedOutcome::RuntimeError(exp), RecordedOutcome::RuntimeError(got)) => Some((
-            format!("runtime_error: {}", exp),
-            format!("runtime_error: {}", got),
-            None,
-        )),
-        (exp, got) => Some((format!("{:?}", exp), format!("{:?}", got), None)),
-    }
-}
+// Per-backend replay drivers + the unified `BackendReplayOutcome`
+// shape they return live in `replay_cmd/backends.rs`. Everything
+// below this line concerns CLI orchestration: file discovery,
+// rendering, the `cmd_replay` entry point.
 
 fn replay_recording_file_vm(
     path: &Path,
@@ -267,97 +170,6 @@ fn replay_recording_file_vm(
         entry_line,
         outcome,
     ))
-}
-
-/// Core VM replay path, returning the backend-agnostic outcome. Wraps
-/// the original "compile, instantiate VM, drive replay" sequence
-/// previously inlined in `replay_recording_file_vm`. Pipeline /
-/// `vm::compile_program_with_modules` errors come back as `Err`
-/// instead of `?` so a single bad recording in `aver replay <dir>`
-/// no longer skips later files via early return.
-fn run_vm_replay(
-    recording: &SessionRecording,
-    replay_module_root: &str,
-    items: &mut Vec<TopLevel>,
-    check_args: bool,
-) -> Result<BackendReplayOutcome, String> {
-    // Full pipeline. Recordings store effect syscalls (`Console.print(...)`
-    // with concrete args) — they don't reference IR shape or bytecode, so
-    // running interp_lower + buffer_build during replay produces the same
-    // effect sequence as the original recording.
-    let pipeline_result = aver::ir::pipeline::run(
-        items,
-        aver::ir::PipelineConfig {
-            typecheck: Some(aver::ir::TypecheckMode::Full {
-                base_dir: Some(replay_module_root),
-            }),
-            ..Default::default()
-        },
-    );
-    let tc_result = pipeline_result.typecheck.expect("typecheck was requested");
-    if !tc_result.errors.is_empty() {
-        return Err(crate::shared::format_type_errors(&tc_result.errors));
-    }
-
-    let mut arena = aver::nan_value::Arena::new();
-    vm::register_service_types(&mut arena);
-    let (code, globals) = vm::compile_program_with_modules(
-        items,
-        &mut arena,
-        Some(replay_module_root),
-        &recording.program_file,
-        pipeline_result.analysis.as_ref(),
-    )
-    .map_err(|e| format!("VM compile error: {}", e))?;
-    let mut machine = vm::VM::new(code, globals, arena);
-    apply_runtime_policy_to_vm(&mut machine, replay_module_root)?;
-    machine.start_replay(recording.effects.clone(), check_args);
-
-    let progress_msg = |machine: &vm::VM, e: &dyn std::fmt::Display| -> String {
-        let (consumed, total) = machine.replay_progress();
-        format!(
-            "Replay failed: {}\nProgress: consumed {} of {} recorded effects",
-            e, consumed, total
-        )
-    };
-
-    machine
-        .run_top_level()
-        .map_err(|e| progress_msg(&machine, &e))?;
-
-    let entry_args = decode_entry_args(&recording.input)?;
-    let nv_args: Vec<NanValue> = entry_args
-        .iter()
-        .map(|v| NanValue::from_value(v, &mut machine.arena))
-        .collect();
-
-    let run_out = machine
-        .run_named_function(&recording.entry_fn, &nv_args)
-        .map_err(|e| progress_msg(&machine, &e))?;
-
-    let actual = if run_out.is_err() {
-        let inner = run_out.wrapper_inner(&machine.arena);
-        RecordedOutcome::RuntimeError(format!(
-            "{} returned error: {}",
-            recording.entry_fn,
-            inner.repr(&machine.arena)
-        ))
-    } else {
-        let val = run_out.to_value(&machine.arena);
-        RecordedOutcome::Value(value_to_json(&val)?)
-    };
-
-    machine
-        .ensure_replay_consumed()
-        .map_err(|e| progress_msg(&machine, &format!("{:?}", e) as &dyn std::fmt::Display))?;
-
-    let (effects_consumed, effects_total) = machine.replay_progress();
-    Ok(BackendReplayOutcome {
-        actual,
-        effects_consumed,
-        effects_total,
-        args_diff_count: machine.args_diff_count(),
-    })
 }
 
 /// Replay a recording through the wasm-gc backend. Phase 4a — wires
@@ -408,54 +220,6 @@ fn replay_recording_file_wasm_gc(
     }
 }
 
-/// Core wasm-gc replay path, returning the backend-agnostic outcome.
-/// Re-materialises the entry call from the recording (default
-/// `main` ↔ no `entry_info`; `--expr` recording ↔ `(fn_name,
-/// decoded_args)`), drives the wasm-gc executor under
-/// `EffectMode::Replaying`, and maps `RunOutcome` onto the shared
-/// shape `build_replay_result` consumes.
-#[cfg(feature = "wasm")]
-fn run_wasm_gc_replay(
-    recording: &SessionRecording,
-    replay_module_root: &str,
-    replay_program_file: &str,
-    check_args: bool,
-) -> Result<BackendReplayOutcome, String> {
-    let entry_info: Option<(String, Vec<Value>)> =
-        if recording.entry_fn == "main" && matches!(&recording.input, JsonValue::Null) {
-            None
-        } else {
-            let args = decode_entry_args(&recording.input)?;
-            Some((recording.entry_fn.clone(), args))
-        };
-    let mode = super::run_wasm_gc::EffectMode::Replaying(Box::new(recording.clone()), check_args);
-    let run = super::run_wasm_gc::try_run_wasm_gc(
-        replay_program_file,
-        Some(replay_module_root),
-        Vec::new(),
-        mode,
-        entry_info,
-    )?;
-    Ok(BackendReplayOutcome {
-        actual: RecordedOutcome::Value(run.output),
-        effects_consumed: run.effects_consumed,
-        effects_total: run.effects_total,
-        args_diff_count: run.args_diff_count,
-    })
-}
-
-/// Look up a function definition's source line by re-parsing the
-/// program file. Best-effort: returns 1 if the file is unreadable or
-/// the function isn't found at module level — same fallback the VM
-/// path uses through its already-parsed AST.
-#[cfg(feature = "wasm")]
-fn find_fn_line_in_file(program_file: &str, name: &str) -> usize {
-    match read_file(program_file).and_then(|src| parse_file(&src)) {
-        Ok(items) => find_fn_line(&items, name),
-        Err(_) => 1,
-    }
-}
-
 fn replay_recording_file_self_host(
     path: &Path,
     _diff: bool,
@@ -483,74 +247,7 @@ fn replay_recording_file_self_host(
     ))
 }
 
-/// Core self-host replay path, returning the backend-agnostic
-/// outcome. Spawns the cached self-host binary against the recording
-/// and treats a successful exit as a full match — the self-host
-/// runtime owns its own `ensure_replay_consumed` check internally and
-/// signals failure via non-zero exit, so a clean exit means the
-/// entire trace was consumed. Output value comparison is not yet
-/// wired here (the binary doesn't surface the entry-fn return as
-/// JSON), so `actual` mirrors `recording.output` to keep the unified
-/// outcome shape honest with what the binary actually verified.
-fn run_self_host_replay(
-    recording: &SessionRecording,
-    replay_module_root: &str,
-    recording_path: &Path,
-    check_args: bool,
-) -> Result<BackendReplayOutcome, String> {
-    let replay_program_file = resolve_replay_program_file(recording, replay_module_root);
-    let binary_path = find_self_host_binary()?;
-    let guest_args = decode_self_host_guest_args(&recording.input)?;
-
-    let mut command = Command::new(&binary_path);
-    command
-        .arg(&replay_program_file)
-        .arg(replay_module_root)
-        .args(&guest_args)
-        .env("AVER_REPLAY_ENTRY_FN", "main")
-        .env("AVER_REPLAY_REPLAY", recording_path)
-        .env("AVER_REPLAY_MODULE_ROOT", replay_module_root)
-        .env_remove("AVER_REPLAY_RECORD")
-        .env_remove("AVER_REPLAY_REQUEST_ID")
-        .env_remove("AVER_REPLAY_TIMESTAMP")
-        .env_remove("AVER_REPLAY_PROGRAM_FILE");
-    if check_args {
-        command.env("AVER_REPLAY_CHECK_ARGS", "1");
-    } else {
-        command.env_remove("AVER_REPLAY_CHECK_ARGS");
-    }
-
-    let output = command.output().map_err(|e| {
-        format!(
-            "Failed to run cached self-host replay binary '{}': {}",
-            binary_path.display(),
-            e
-        )
-    })?;
-
-    if !output.status.success() {
-        let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
-        let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-        let mut msg = "Self-host replay failed".to_string();
-        if !stdout.is_empty() {
-            msg.push_str(&format!("\nstdout:\n{}", stdout));
-        }
-        if !stderr.is_empty() {
-            msg.push_str(&format!("\nstderr:\n{}", stderr));
-        }
-        return Err(msg);
-    }
-
-    let n = recording.effects.len();
-    Ok(BackendReplayOutcome {
-        actual: recording.output.clone(),
-        effects_consumed: n,
-        effects_total: n,
-        args_diff_count: 0,
-    })
-}
-
-enum ReplayError {
+pub(super) enum ReplayError {
     Generic(String),
 }
 
@@ -795,7 +492,8 @@ pub(super) fn cmd_replay(
 
 #[cfg(test)]
 mod tests {
-    use super::{collect_recording_files, decode_self_host_guest_args};
+    use super::backends::decode_self_host_guest_args;
+    use super::collect_recording_files;
     use aver::replay::JsonValue;
     use std::fs;
     use std::path::PathBuf;
